Some refactorization
