+++ src/app/page/input/input-pane/input-pane.component.css
@@ -4,7 +4,6 @@
 
 :host > app-time-unit-selector {
     width: 100%;
-    height: 102px;
 }
 
 :host > div {
