+++ src/app/page/input/input-pane/input-pane.component.html
