+++ src/app/page/input/tabular-input/tabular-input.component.css
