+++ src/app/page/input/time-unit-selector/time-unit-selector.component.css
@@ -1,0 +1,15 @@
+:host {
+    display: block;
+}
+
+:host > div > label,
+:host > div > select {
+    font-size: 14px;
+}
+
+@media only screen and (min-width: 800px) {
+    :host > div > label,
+    :host > div > select {
+        font-size: 16px;
+    }
+}
