+++ src/app/page/output/chains-result/chains-result.component.ts
