+++ src/app/page/output/tabular-output/tabular-output.component.css
@@ -1,0 +1,3 @@
+:host {
+    display: block;
+}
