+++ src/app/page/output/tabular-output/tabular-output.component.ts
